The trimming and native AOT story is a how-to, not a csproj comment
What Quartz claims about trimming lived in a 20-line comment in src/Quartz/Quartz.csproj
and in the remarks on two build targets, so an application that wanted a trimmed or native
publish had to work out from build errors which parts of Quartz are reflective.

docs/documentation/quartz-4.x/how-tos/trimming-and-native-aot.md is that, as a page: what
IsTrimmable and IsAotCompatible each claim and what neither does, the string-named contracts
that still warn with an action for each, the recipe (a DbProviderFactory rather than a driver
name, job types named as types, job-data value types through AddTypeInfoResolver), the canary
as the worked example, the two packages that are deliberately not trimmable, and where the
rest of the category stands.

The tutorial's four trimming sections become two paragraphs that state the claim and link
here; the packages table moves rather than being copied, and the migration guide's links
follow it.

Three claims the seed made are corrected against the code and the ILLink sources rather
than repeated: an unmarked assembly is not kept whole under TrimMode=full, the IL2104
collapse is TrimmerSingleWarn rather than IsTrimmable, and a registered DbDataSource is not
an equally name-free registration.
